rego-compile: utility->policytemplates, /ap1/1.0 prefix duplication removal
