+++ api/swagger/docs.go
@@ -214,6 +214,529 @@ const docTemplate = `{
                 }
             }
         },
+        "/admin/policytemplates": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "정책 템플릿 목록을 조회한다. 정책 템플릿 목록 조회 결과는 최신 템플릿 버전 목록만 조회된다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[ListPolicyTemplate] 정책 템플릿 목록 조회",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "pageSize",
+                        "name": "limit",
+                        "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "pageNumber",
+                        "name": "page",
+                        "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "sortColumn",
+                        "name": "soertColumn",
+                        "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "sortOrder",
+                        "name": "sortOrder",
+                        "in": "query"
+                    },
+                    {
+                        "type": "array",
+                        "items": {
+                            "type": "string"
+                        },
+                        "collectionFormat": "csv",
+                        "description": "filters",
+                        "name": "filters",
+                        "in": "query"
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.ListPolicyTemplateResponse"
+                        }
+                    }
+                }
+            },
+            "post": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "정책 템플릿을 신규 생성(v1.0.0을 생성)한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[CreatePolicyTemplate] 정책 템플릿 신규 생성",
+                "parameters": [
+                    {
+                        "description": "create policy template request",
+                        "name": "body",
+                        "in": "body",
+                        "required": true,
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreatePolicyTemplateRequest"
+                        }
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreatePolicyTemplateReponse"
+                        }
+                    }
+                }
+            }
+        },
+        "/admin/policytemplates/kind/{policyTemplateKind}/existence": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 유형을 가진 정책 템플릿이 이미 존재하는지 확인한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[ExistsPolicyTemplateKind] 정책 템플릿 유형 존재 여부 확인",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 이름",
+                        "name": "policyTemplateKind",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.ExistsPolicyTemplateKindResponse"
+                        }
+                    }
+                }
+            }
+        },
+        "/admin/policytemplates/name/{policyTemplateName}/existence": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 이름을 가진 정책 템플릿이 이미 존재하는지 확인한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[ExistsPolicyTemplateName] 정책 템플릿 아름 존재 여부 확인",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 이름",
+                        "name": "policyTemplateName",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CheckExistedResponse"
+                        }
+                    }
+                }
+            }
+        },
+        "/admin/policytemplates/{policyTemplateId}": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 식별자를 가진 정책 템플릿의 최신 버전을 조회한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[GetPolicyTemplate] 정책 템플릿 조회(최신 버전)",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetPolicyTemplateResponse"
+                        }
+                    }
+                }
+            },
+            "delete": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "정책 템플릿을 삭제한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[DeletePolicyTemplate] 정책 템플릿 삭제",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK"
+                    }
+                }
+            },
+            "patch": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "정책 템플릿의 업데이트 가능한 필드들을 업데이트한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[UpdatePolicyTemplate] 정책 템플릿 업데이트",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    },
+                    {
+                        "description": "update  policy template request",
+                        "name": "body",
+                        "in": "body",
+                        "required": true,
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.UpdatePolicyTemplateRequest"
+                        }
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK"
+                    }
+                }
+            }
+        },
+        "/admin/policytemplates/{policyTemplateId}/deploy": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 식별자를 가진 정책 템플릿의 정책 템플릿 클러스터 별 설치 버전을 조회한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[GetPolicyTemplateDeploy] 정책 템플릿 클러스터 별 설치 버전 조회",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetPolicyTemplateDeployResponse"
+                        }
+                    }
+                }
+            }
+        },
+        "/admin/policytemplates/{policyTemplateId}/statistics": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 식별자를 가진 정책 템플릿의 최신 버전을 조회한다. 전체 조직의 통계를 조회하려면 organizationId를 tks로 설정한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[ListPolicyTemplateStatistics] 정책 템플릿 사용 카운트 조회",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.ListPolicyTemplateStatisticsResponse"
+                        }
+                    }
+                }
+            }
+        },
+        "/admin/policytemplates/{policyTemplateId}/versions": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 식별자를 가진 정책 템플릿의 최신 버전을 조회한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[ListPolicyTemplateVersions] 정책 템플릿 버전목록 조회",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.ListPolicyTemplateVersionsResponse"
+                        }
+                    }
+                }
+            },
+            "post": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 식별자를 가진 정책 템플릿의 특정 버전을 저장한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[CreatePolicyTemplateVersion] 정책 템플릿 특정 버전 저장",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    },
+                    {
+                        "description": "create policy template version request",
+                        "name": "body",
+                        "in": "body",
+                        "required": true,
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreatePolicyTemplateVersionRequest"
+                        }
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreatePolicyTemplateVersionResponse"
+                        }
+                    }
+                }
+            }
+        },
+        "/admin/policytemplates/{policyTemplateId}/versions/{version}": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 식별자를 가진 정책 템플릿의 특정 버전을 조회한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[GetPolicyTemplateVersion] 정책 템플릿 특정 버전 조회",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "조회할 버전(v0.0.0 형식)",
+                        "name": "version",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetPolicyTemplateVersionResponse"
+                        }
+                    }
+                }
+            },
+            "delete": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "해당 식별자를 가진 정책 템플릿의 특정 버전을 삭제한다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[DeletePolicyTemplateVersion] 정책 템플릿 특정 버전 삭제",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "정책 템플릿 식별자(uuid)",
+                        "name": "policyTemplateId",
+                        "in": "path",
+                        "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "삭제할 버전(v0.0.0 형식)",
+                        "name": "version",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK"
+                    }
+                }
+            }
+        },
         "/admin/stack-templates": {
             "get": {
                 "security": [
@@ -473,529 +996,6 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.UpdateStackTemplateOrganizationsRequest"
                         }
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK"
-                    }
-                }
-            }
-        },
-        "/api/1.0/admin/policytemplates": {
-            "get": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "정책 템플릿 목록을 조회한다. 정책 템플릿 목록 조회 결과는 최신 템플릿 버전 목록만 조회된다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[ListPolicyTemplate] 정책 템플릿 목록 조회",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "pageSize",
-                        "name": "limit",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "pageNumber",
-                        "name": "page",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "sortColumn",
-                        "name": "soertColumn",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "sortOrder",
-                        "name": "sortOrder",
-                        "in": "query"
-                    },
-                    {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "collectionFormat": "csv",
-                        "description": "filters",
-                        "name": "filters",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.ListPolicyTemplateResponse"
-                        }
-                    }
-                }
-            },
-            "post": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "정책 템플릿을 신규 생성(v1.0.0을 생성)한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[CreatePolicyTemplate] 정책 템플릿 신규 생성",
-                "parameters": [
-                    {
-                        "description": "create policy template request",
-                        "name": "body",
-                        "in": "body",
-                        "required": true,
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreatePolicyTemplateRequest"
-                        }
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreatePolicyTemplateReponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/api/1.0/admin/policytemplates/kind/{policyTemplateKind}/existence": {
-            "get": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 유형을 가진 정책 템플릿이 이미 존재하는지 확인한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[ExistsPolicyTemplateKind] 정책 템플릿 유형 존재 여부 확인",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 이름",
-                        "name": "policyTemplateKind",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.ExistsPolicyTemplateKindResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/api/1.0/admin/policytemplates/name/{policyTemplateName}/existence": {
-            "get": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 이름을 가진 정책 템플릿이 이미 존재하는지 확인한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[ExistsPolicyTemplateName] 정책 템플릿 아름 존재 여부 확인",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 이름",
-                        "name": "policyTemplateName",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CheckExistedResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/api/1.0/admin/policytemplates/{policyTemplateId}": {
-            "get": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 식별자를 가진 정책 템플릿의 최신 버전을 조회한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[GetPolicyTemplate] 정책 템플릿 조회(최신 버전)",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetPolicyTemplateResponse"
-                        }
-                    }
-                }
-            },
-            "delete": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "정책 템플릿을 삭제한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[DeletePolicyTemplate] 정책 템플릿 삭제",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK"
-                    }
-                }
-            },
-            "patch": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "정책 템플릿의 업데이트 가능한 필드들을 업데이트한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[UpdatePolicyTemplate] 정책 템플릿 업데이트",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "description": "update  policy template request",
-                        "name": "body",
-                        "in": "body",
-                        "required": true,
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.UpdatePolicyTemplateRequest"
-                        }
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK"
-                    }
-                }
-            }
-        },
-        "/api/1.0/admin/policytemplates/{policyTemplateId}/deploy": {
-            "get": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 식별자를 가진 정책 템플릿의 정책 템플릿 클러스터 별 설치 버전을 조회한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[GetPolicyTemplateDeploy] 정책 템플릿 클러스터 별 설치 버전 조회",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetPolicyTemplateDeployResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/api/1.0/admin/policytemplates/{policyTemplateId}/statistics": {
-            "get": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 식별자를 가진 정책 템플릿의 최신 버전을 조회한다. 전체 조직의 통계를 조회하려면 organizationId를 tks로 설정한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[ListPolicyTemplateStatistics] 정책 템플릿 사용 카운트 조회",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.ListPolicyTemplateStatisticsResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/api/1.0/admin/policytemplates/{policyTemplateId}/versions": {
-            "get": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 식별자를 가진 정책 템플릿의 최신 버전을 조회한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[ListPolicyTemplateVersions] 정책 템플릿 버전목록 조회",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.ListPolicyTemplateVersionsResponse"
-                        }
-                    }
-                }
-            },
-            "post": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 식별자를 가진 정책 템플릿의 특정 버전을 저장한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[CreatePolicyTemplateVersion] 정책 템플릿 특정 버전 저장",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "description": "create policy template version request",
-                        "name": "body",
-                        "in": "body",
-                        "required": true,
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreatePolicyTemplateVersionRequest"
-                        }
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreatePolicyTemplateVersionResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/api/1.0/admin/policytemplates/{policyTemplateId}/versions/{version}": {
-            "get": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 식별자를 가진 정책 템플릿의 특정 버전을 조회한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[GetPolicyTemplateVersion] 정책 템플릿 특정 버전 조회",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "string",
-                        "description": "조회할 버전(v0.0.0 형식)",
-                        "name": "version",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetPolicyTemplateVersionResponse"
-                        }
-                    }
-                }
-            },
-            "delete": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "해당 식별자를 가진 정책 템플릿의 특정 버전을 삭제한다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "PolicyTemplate"
-                ],
-                "summary": "[DeletePolicyTemplateVersion] 정책 템플릿 특정 버전 삭제",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "정책 템플릿 식별자(uuid)",
-                        "name": "policyTemplateId",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "string",
-                        "description": "삭제할 버전(v0.0.0 형식)",
-                        "name": "version",
-                        "in": "path",
-                        "required": true
                     }
                 ],
                 "responses": {
@@ -6538,6 +6538,52 @@ const docTemplate = `{
                 }
             }
         },
+        "/policytemplates/rego-compile": {
+            "post": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "Rego 코드 컴파일 및 파라미터 파싱을 수행한다. 파라미터 파싱을 위해서는 먼저 컴파일이 성공해야 하며, parseParameter를 false로 하면 컴파일만 수행할 수 있다.",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "PolicyTemplate"
+                ],
+                "summary": "[CompileRego] Rego 코드 컴파일 및 파라미터 파싱",
+                "parameters": [
+                    {
+                        "type": "boolean",
+                        "description": "파라미터 파싱 여부",
+                        "name": "parseParameter",
+                        "in": "query",
+                        "required": true
+                    },
+                    {
+                        "description": "Rego 코드",
+                        "name": "body",
+                        "in": "body",
+                        "required": true,
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.RegoCompileRequest"
+                        }
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.RegoCompileResponse"
+                        }
+                    }
+                }
+            }
+        },
         "/system-api/organizations/{organizationId}/alerts": {
             "post": {
                 "security": [
@@ -6568,52 +6614,6 @@ const docTemplate = `{
                 "responses": {
                     "200": {
                         "description": "OK"
-                    }
-                }
-            }
-        },
-        "/utility/rego-compile": {
-            "post": {
-                "security": [
-                    {
-                        "JWT": []
-                    }
-                ],
-                "description": "Rego 코드 컴파일 및 파라미터 파싱을 수행한다. 파라미터 파싱을 위해서는 먼저 컴파일이 성공해야 하며, parseParameter를 false로 하면 컴파일만 수행할 수 있다.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "Rego"
-                ],
-                "summary": "[CompileRego] Rego 코드 컴파일 및 파라미터 파싱",
-                "parameters": [
-                    {
-                        "type": "boolean",
-                        "description": "파라미터 파싱 여부",
-                        "name": "parseParameter",
-                        "in": "query",
-                        "required": true
-                    },
-                    {
-                        "description": "Rego 코드",
-                        "name": "body",
-                        "in": "body",
-                        "required": true,
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.RegoCompileRequest"
-                        }
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.RegoCompileResponse"
-                        }
                     }
                 }
             }
